requirements.txt
# CUDA 11.8
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
torchrl
tensorboard
gymnasium == 0.29.1
mujoco[all]
genesis
pygame
matplotlib
numpy
pandas
hydra-core
